remove silent_push option
the silent_push option was meant to be a way of choosing wheather to
send APNS silent notifications or not. After experimenting with APNS
push notifications it became clear that silent notifications are the better
alternative for xmpp clients. They are more generic and the throttling
problem I read somewhere about seems to not exist.