notify previous users
After a server restart previous pending users will now get sent a push
notification so they can wake up and start a new connection. Currently
the notification is not distinguishable from notifications sent in other
events. So a client will typically try to resume the stream, which fails
because the stream does not exist anymore. After that it will open a new
stream and re-enable push.