add retry mechanism and error handling to mod_push_apns
The APNS backend now implements error handling and retrying. An error is
recoverable if APNS responds with status code 10 (Shutdown) or 7
(invalid payload size). The retry period is set by ?RETRY_INTERVAL, the
time the backend waits for an error response is set by
?PENDING_INTERVAL. We will try to send multiple notifications at once,
the maximum number is set by ?MAX_PENDING_NOTIFICATIONS.
The ciphersuites had to be changed as Apple apparently changed their TLS
configuration. This needs more investigation.