error handling fix, add retry_timestamp
When APNS reports an error the respective notification and notifications
sent earlier must be removed from the pending queue. This is now done
correctly.
As with the pending_timer we now check for the start timestamp in the
retry handler. This is done for the unlikely event that retry_timer
times out before a handler tries to stop / restart it