README: remove "for now" qualifier on ejabberd requirements
This creates a lot of ambiguity around when "now" actually is or was,
whether it's `still true', etc. And it leads people to waste
both their own time and others' trying to determine
whether those changes have been integrated into the mainline ejabberd
in the intervening years between when the "for now" statement was first made
and the release of ejabberd that they're running, or whether
the statement is still actually current (which it is).